Start converting AppContext into something for this app
Involed a lot of trial and error on the listeners, because the listener
sizes stop being known at runtime when things start getting dispersed
into modules.
