+++ gtk/src/components/mod.rs
@@ -1,5 +1,8 @@
+use gtk::prelude::*;
+
 mod basics;
 mod day;
+mod history;
 mod rep_duration;
 mod set_rep;
 mod steps;
@@ -7,3 +10,4 @@ mod time_distance;
 mod weight;
 
 pub use day::day_c;
+pub use history::History;
